Initial commit
This is super hacked together while I was learning the Google calendar
APIs and I've yet to really pick it apart and break it down into
separate modules, and I honestly don't know if I intend to. Making a
repo for it so I can start working on it at work without having to use
my flash drive to move it around.