Clean up EventBuilder
Changed the scoping in getLength. Pretty much the entirety of the method was pasted from the original script from before I really knew what I was doing with node and javascript. Though, I don't actually think const or let were around back then.

I'm sure I'll work out a better way to handle this method overall in the future but for now it works as it did originally and it gets the job done